msgbroker_test.go
package ratebroker

import (
	"context"
	"errors"
	"testing"
	"time"
)

func TestRedisMessageBroker_Publish(t *testing.T) {
	// This is a simplified example of what your RateEvent might look like.
	sampleMessage := RateEvent{ /* fields */ }

	t.Run("publishes message successfully", func(t *testing.T) {
		broker := RedisMessageBroker{
			// initialize with a buffered channel to simulate successful send
			publishChannel: make(chan RateEvent, 1),
			// other necessary initializations...
		}

		// Don't forget to close channels if your implementation doesn't close them.
		defer close(broker.publishChannel)

		ctx := context.Background() // using a background context with no deadline for simplicity
		err := broker.Publish(ctx, sampleMessage)

		if err != nil {
			t.Errorf("expected no error, got %v", err)
		}
	})

	t.Run("fails to publish due to context done", func(t *testing.T) {
		broker := RedisMessageBroker{
			publishChannel: make(chan RateEvent),
			// other necessary initializations...
		}

		// Don't forget to close channels if your implementation doesn't close them.
		defer close(broker.publishChannel)

		ctx, cancel := context.WithCancel(context.Background())

		// cancel the context to simulate a closed/deadline-exceeded context
		cancel()

		err := broker.Publish(ctx, sampleMessage)

		if !errors.Is(err, context.Canceled) {
			t.Errorf("expected context.Canceled error, got %v", err)
		}
	})

	t.Run("fails to publish due to full channel", func(t *testing.T) {
		broker := RedisMessageBroker{
			// initialize with an unbuffered channel to simulate a full channel scenario
			publishChannel: make(chan RateEvent),
			// other necessary initializations...
		}

		// Note: Not deferring close on the channel because it should block and not accept the message.

		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
		defer cancel() // always clean up resources

		// This Goroutine tries to read from the channel to ensure there's no deadlock
		// if the implementation of Publish changes. However, it does not read anything
		// so the channel remains full for the test case.
		go func() {
			select {
			case <-ctx.Done():
				// Context completed before message was received (expected because we're testing the full channel)
			case <-broker.publishChannel:
				// Not expecting to receive anything
			}
		}()

		err := broker.Publish(ctx, sampleMessage)

		// We're expecting a default case error due to the select in the Publish method.
		if err == nil || err.Error() != "publish operation could not proceed; channel full" {
			t.Errorf("expected 'channel full' error, got %v", err)
		}
	})
}
